\name{dist-class}
\docType{class}
\alias{dist-class}

\title{Class "dist"}
\description{Class to represent distribution functions.}
\section{Objects from the Class}{
This is a virtual class. Create objects of drived classes instead.
}
\section{Methods}{
No methods defined with class "dist" in the signature.
}
\author{Peter Humburg}

\seealso{
	\code{\linkS4class{discDist}}, \code{\linkS4class{contDist}}, \code{\linkS4class{tDist}}
}
\examples{
showClass("dist")
}
\keyword{classes}
\keyword{distribution}